requirements.txt
mcp>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.0.0
